+++ changelog.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [2.3.0] - 2020-03-02
+
+### Added
+
+- Rolled all add-ons into core.
+
+### Changed
+
+- Placed all Members-related settings under one admin menu item.
+- Made login and user widgets enabled by default, and removed settings.
+
 ## [2.2.0] - 2019-07-20
 
 ### Added